conversion: link-edit.js to tsx
